requirements.txt
streamlit
streamlit-drawable-canvas
fastapi
uvicorn
pillow
torch
torchvision
transformers

